1.Добавлены ответы с сервера 2.Функции в JS обработывают ответы с ручки и проверяют их. 3.CRUD готов.

+++ src/main/java/application/controller/UserController.java
@@ -28,67 +28,40 @@ public class UserController {
         this.userService = userService;
     }
 
-    //    @PreAuthorize("hasRole('ROLE_ADMIN')")
-//    @GetMapping("/admin")
-//    public String getListOfUsers(Model model) {
-//        List<User> userList = userService.findAll();
-//        model.addAttribute("user", userList);
-//        return "user_list";
-//    }
 //    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @GetMapping("user-list-json")
     public List<User> getListOfUsers() {
         List<User> userList = userService.findAll();
+        System.out.println(userList);
         return userList;
     }
 
-    @GetMapping("admin/edit/{id}")
-    public String editUser(Model model, @PathVariable("id") long id) {
-        User user = userService.findById(id);
-        model.addAttribute("user", user);
-        return "update_user";
-    }
-
-    @PostMapping("admin/edit/update/{id}")
-    public String editUser(@PathVariable("id") long id, User user, Model model) {
-        userService.save(user);
-        return "redirect:/admin";
-    }
-
     @RequestMapping(value = "user-post-json", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public void deleteUserJson(@RequestBody User user) {
-        userService.deleteUserById(user.getId());
+    public ResponseEntity<User> deleteUserJson(@RequestBody User user) {
+        if (user != null) {
+            userService.deleteUserById(user.getId());
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @RequestMapping(value = "user-save-json", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public void saveUserJson(@RequestBody User user) throws JsonProcessingException {
-        userService.save(user);
+    public ResponseEntity<User> saveUserJson(@RequestBody User user) throws JsonProcessingException {
+        if (user != null) {
+            userService.save(user);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
-    @GetMapping("admin/delete/{id}")
-    public String deleteUser(Model model, @PathVariable("id") long id) {
-        userService.deleteUserById(id);
-        return "redirect:/admin";
-    }
-
-    @GetMapping("user")
-    public String getCurrentUser(Model model) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        User user = (User) authentication.getPrincipal();
-        model.addAttribute("user", user);
-        return "user_info";
-    }
-
-
-    @GetMapping("admin/addUser")
-    public String addUser(User user) {
-        return "add_user";
-    }
-
-    @PostMapping("admin/saveUser")
-    public String saveUser(User user) {
-        userService.save(user);
-        return "redirect:/admin";
-    }
+//    @GetMapping("user")
+//    public String getCurrentUser(Model model) {
+//        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//        User user = (User) authentication.getPrincipal();
+//        model.addAttribute("user", user);
+//        return "user_info";
+//    }
 
 }
